improved button functionality
